Update Absolute Schedule: Fix some bugs, rename AbsoluteScheduler back to Scheduler
