+++ src/components/shared/Scheduler.vue
@@ -4,18 +4,15 @@ Prevention -- A research institute of the Ludwig Boltzmann Gesellschaft,
 Oesterreichische Vereinigung zur Foerderung der wissenschaftlichen Forschung).
 Licensed under the Elastic License 2.0. */
 <script setup lang="ts">
-  import { inject, ref, Ref } from 'vue';
+  import { inject, ref, Ref, watch } from 'vue';
   import Calendar from 'primevue/calendar';
   import Button from 'primevue/button';
-  import InputText from 'primevue/inputtext';
-  import Dropdown from 'primevue/dropdown';
-  import SelectButton from 'primevue/selectbutton';
   import Checkbox from 'primevue/checkbox';
-  import { Frequency, Weekday, Event } from '../../generated-sources/openapi';
-  import { Nullable } from 'vitest';
+  import { Event, RecurrenceRule } from '../../generated-sources/openapi';
   import { useI18n } from 'vue-i18n';
   import { useStudyStore } from '../../stores/studyStore';
-  import { dateToDateTimeString } from '../../utils/dateUtils';
+  import { MoreTableChoice } from '../../models/MoreTableModel';
+  import AbsoluteSchedulerRepetition from '../subComponents/AbsoluteSchedulerRepetition.vue';
 
   const { t } = useI18n();
   const dialogRef: any = inject('dialogRef');
@@ -23,478 +20,188 @@ Licensed under the Elastic License 2.0. */
   const studyStore = useStudyStore();
 
   const scheduler: any = dialogRef.value.data.scheduler;
+  const returnSchedule: Ref<Event> = ref({
+    type: scheduler.type ? scheduler.type : 'Event',
+    dtstart: scheduler.dtstart
+      ? scheduler.dtstart
+      : studyStore.study.plannedStart,
+    dtend: scheduler.dtend ? scheduler.dtend : studyStore.study.plannedEnd,
+    rrule: scheduler.rrule,
+  });
 
-  const repetitionEndArray = [
-    {
-      label: t('scheduler.labels.event.repetitionEnd.studyEnd'),
-      value: 'onDate',
-      active: true,
-      unit: 'onDate',
-    },
-    {
-      label: t('scheduler.labels.event.repetitionEnd.after'),
-      value: 'after',
-      active: true,
-      unit: 'after',
-    },
-  ];
-
-  const repeatFreqArray = [
-    {
-      label: t('scheduler.frequency.labels.daily'),
-      value: Frequency.Daily,
-      active: true,
-      unit: t('scheduler.frequency.days'),
-    },
-    {
-      label: t('scheduler.frequency.labels.weekly'),
-      value: Frequency.Weekly,
-      active: true,
-      unit: t('scheduler.frequency.weeks'),
-    },
-    //{label: 'Monthly', value: Frequency.Monthly, active: true, unit: 'month(s)'}
-  ];
-  const repeatWeekdayArray = [
-    { label: t('scheduler.weekday.short.monday'), value: Weekday.Mo },
-    { label: t('scheduler.weekday.short.tuesday'), value: Weekday.Tu },
-    { label: t('scheduler.weekday.short.wednesday'), value: Weekday.We },
-    { label: t('scheduler.weekday.short.thursday'), value: Weekday.Th },
-    { label: t('scheduler.weekday.short.friday'), value: Weekday.Fr },
-    { label: t('scheduler.weekday.short.saturday'), value: Weekday.Sa },
-    { label: t('scheduler.weekday.short.sunday'), value: Weekday.Su },
-  ];
-  const repeatYearOptionArray = [
-    { label: t('scheduler.labels.selection.specificDay'), value: 'onSpecific' },
-    //{label: 'On specific options', value: 'onOptions'}
-  ];
-  const repeatBySetPosOptionArray = [
-    { label: t('scheduler.bySetPosLabel.first'), value: 1 },
-    { label: t('scheduler.bySetPosLabel.second'), value: 2 },
-    { label: t('scheduler.bySetPosLabel.third'), value: 3 },
-    { label: t('scheduler.bySetPosLabel.fourth'), value: 4 },
-    { label: t('scheduler.bySetPosLabel.last'), value: -1 },
-  ];
-  const repeatByMonthOptionArray = [
-    { label: t('scheduler.months.short.january'), value: 1 },
-    { label: t('scheduler.months.short.january'), value: 2 },
-    { label: t('scheduler.months.short.january'), value: 3 },
-    { label: t('scheduler.months.short.january'), value: 4 },
-    { label: t('scheduler.months.short.january'), value: 5 },
-    { label: t('scheduler.months.short.january'), value: 6 },
-    { label: t('scheduler.months.short.january'), value: 7 },
-    { label: t('scheduler.months.short.january'), value: 8 },
-    { label: t('scheduler.months.short.january'), value: 9 },
-    { label: t('scheduler.months.short.january'), value: 10 },
-    { label: t('scheduler.months.short.january'), value: 11 },
-    { label: t('scheduler.months.short.january'), value: 12 },
-  ];
-  const repeatByDayForYearArray = [
-    { label: t('scheduler.weekday.monday'), value: [Weekday.Mo] },
-    { label: t('scheduler.weekday.tuesday'), value: [Weekday.Tu] },
-    { label: t('scheduler.weekday.wednesday'), value: [Weekday.We] },
-    { label: t('scheduler.weekday.thursday'), value: [Weekday.Th] },
-    { label: t('scheduler.weekday.friday'), value: [Weekday.Fr] },
-    { label: t('scheduler.weekday.saturday'), value: [Weekday.Sa] },
-    { label: t('scheduler.weekday.sunday'), value: [Weekday.Su] },
-    {
-      label: t('scheduler.labels.selection.allDaysInWeek'),
-      value: [
-        Weekday.Mo,
-        Weekday.Tu,
-        Weekday.We,
-        Weekday.Th,
-        Weekday.Fr,
-        Weekday.Sa,
-        Weekday.Su,
-      ],
-    },
-    {
-      label: t('scheduler.labels.selection.allWeekdays'),
-      value: [Weekday.Mo, Weekday.Tu, Weekday.We, Weekday.Th, Weekday.Fr],
-    },
-    {
-      label: t('scheduler.labels.selection.allWeekendDays'),
-      value: [Weekday.Sa, Weekday.Su],
-    },
-  ];
-
-  const minDate: Date = new Date(studyStore.study.plannedStart as string);
-  minDate.setHours(0, 0, 0);
-  const maxDate: Date = new Date(studyStore.study.plannedEnd as string);
-  maxDate.setHours(23, 59, 59);
-
-  const start: Ref<Date> = ref(
-    scheduler.dtstart ? new Date(scheduler.dtstart) : new Date()
+  const calendarStart: Ref<Date> = ref(
+    new Date(returnSchedule.value.dtstart as string)
   );
-  const end: Ref<Date> = ref(
-    scheduler.dtend ? new Date(scheduler.dtend) : new Date()
+  const calendarEnd: Ref<Date> = ref(
+    new Date(returnSchedule.value.dtend as string)
   );
+  const calendarEndChangedWithStart: Ref<boolean> = ref(false);
 
-  const formerStart: Ref<Date> = ref(start.value);
-  const formerEnd: Ref<Date> = ref(end.value);
+  watch(calendarStart, async (newValue, oldValue) => {
+    validateDate('calendarStart', newValue, oldValue);
+  });
 
-  const allDayChecked: Ref<boolean> = ref(false);
-  const oneDayObservationChecked: Ref<boolean> = ref(true);
-  const repeatChecked: Ref<boolean> = ref(false);
+  watch(calendarEnd, async (newValue, oldValue) => {
+    if (!calendarEndChangedWithStart.value) {
+      console.error('enter calendarEnd Watcher validate Date');
+      validateDate('calendarEnd', newValue, oldValue);
+    }
+  });
 
+  const entireDayCheckbox: Ref<boolean> = ref(false);
   if (
-    scheduler?.dtstart &&
-    scheduler?.dtend &&
-    dateToDateTimeString(new Date(scheduler?.dtstart))?.substring(11, 16) ===
-      '00:00' &&
-    dateToDateTimeString(new Date(scheduler?.dtend))?.substring(11, 16) ===
-      '23:59'
+    typeof scheduler.dtstart !== 'undefined' &&
+    typeof scheduler.dtend !== 'undefined'
   ) {
-    allDayChecked.value = true;
-  }
-
-  if (scheduler && scheduler.dtstart && scheduler.dtend) {
     if (
-      dateToDateTimeString(new Date(scheduler?.dtstart))
-        ?.toString()
-        .substring(0, 10) !==
-      dateToDateTimeString(new Date(scheduler?.dtend))
-        ?.toString()
-        .substring(0, 10)
+      scheduler.dtstart.getHours() === '00' &&
+      scheduler.dtstart.getMinutes() === '00' &&
+      scheduler.dtstart.getSeconds() === '00' &&
+      scheduler.dtend.getHours() === '23' &&
+      scheduler.dtend.getMinutes() === '59' &&
+      scheduler.dtend.getSeconds() === '59'
     ) {
-      oneDayObservationChecked.value = false;
-    } else if (
-      dateToDateTimeString(new Date(scheduler?.dtstart))
-        ?.toString()
-        .substring(0, 10) ===
-      dateToDateTimeString(new Date(scheduler?.dtend))
-        ?.toString()
-        .substring(0, 10)
-    ) {
-      oneDayObservationChecked.value = true;
+      entireDayCheckbox.value = true;
     }
   }
 
-  const repeatFreq: Ref<Frequency | undefined> = ref(
-    scheduler.rrule && scheduler.rrule.freq ? scheduler.rrule.freq : undefined
-  );
-  const repeatInterval: Ref<number | undefined> = ref(
-    scheduler.rrule && scheduler.rrule.interval
-      ? scheduler.rrule.interval
-      : undefined
-  ); // hourly/daily/weekly/monthly/yearly
-  const repeatCount: Ref<Nullable<string>> = ref(
-    scheduler.rrule && scheduler.rrule.count ? scheduler.rrule.count : undefined
-  );
-  // kein repeatUntil wenn repeatCount // hourly/daily/weekly/monthly/yearly
-  const repeatUntil: Ref<Date | undefined> = ref(
-    scheduler.rrule && scheduler.rrule.until
-      ? new Date(scheduler.rrule.until)
-      : undefined
-  ); // kein repeatCount wenn repeatUntil //hourly/daily/weekly/monthly/yearly
-  const repeatByDay: Ref<Weekday[] | undefined> = ref(
-    scheduler.rrule && scheduler.rrule.byday ? scheduler.rrule.byday : undefined
-  ); // weekly/monthly/yearly
-  const repeatByMonth: Ref<number | undefined> = ref(
-    scheduler.rrule && scheduler.rrule.bymonth
-      ? scheduler.rrule.bymonth
-      : undefined
-  ); // monthly/yearly
-  const repeatByMonthDay: Ref<Nullable<string>> = ref(
-    scheduler.rrule && scheduler.rrule.bymonthday
-      ? scheduler.rrule.bymonthday
-      : undefined
-  ); // yearly
-  const repeatBySetPos: Ref<number | undefined> = ref(
-    scheduler.rrule && scheduler.rrule.bysetpos
-      ? scheduler.rrule.bysetpos
-      : undefined
-  ); // monthly/yearly
+  const singleDayEventCheckbox: Ref<boolean> = ref(
+    calendarStart.value.getDay() === calendarEnd.value.getDay() ? true : false
+  ); // Individual Observation Checkbox
+  const hasRruleValue: Ref<boolean> = ref(
+    !returnSchedule.value.rrule ? false : true
+  ); // Repeat Event Checkbox
 
-  const repeatCountLabel: Ref<string | undefined> = ref(
-    scheduler.rrule && scheduler.rrule.freq
-      ? repeatFreqArray.find((f: any) => f.value === scheduler.rrule.freq)?.unit
-      : ''
-  );
-  const repeatYearOption: Ref<string> = ref('onSpecific');
-  const repeatEndOption: Ref<string> = ref('never');
-  //const intervalError: Ref<string> = ref('')
-
-  if (repeatFreq.value) {
-    repeatChecked.value = true;
-  }
-
-  if (repeatCount.value && repeatByDay.value?.length) {
-    repeatCount.value = (
-      parseInt(repeatCount.value) / repeatByDay.value?.length || 1
-    ).toString();
-  }
-
-  if (repeatCount.value || repeatUntil.value) {
-    if (repeatCount.value) {
-      repeatEndOption.value = 'after';
-    }
-    if (repeatUntil.value) {
-      repeatEndOption.value = 'onDate';
-    }
-  }
-
-  if (scheduler && scheduler.rrule) {
+  function onChangeSingleDayEventCheckbox() {
     if (
-      scheduler.rrule.bymonthday ||
-      scheduler.rrule.bysetpos ||
-      scheduler.rrule.byday
+      singleDayEventCheckbox.value &&
+      calendarStart.value &&
+      calendarEnd.value
     ) {
-      if (scheduler.rrule.bymonthday) {
-        repeatYearOption.value = 'onSpecific';
-      }
-      if (scheduler.rrule.bysetpos && scheduler.rrule.byday) {
-        repeatYearOption.value = 'onOptions';
-      }
-    }
-  }
-
-  function setRepeatCountLabel(repeatFreq: string | undefined) {
-    if (repeatFreq) {
-      repeatCountLabel.value = repeatFreqArray.find(
-        (f: any) => f.value === repeatFreq
-      )?.unit;
-    } else {
-      return '';
-    }
-  }
-
-  function setRepetitionEnd(repeatEnd: string | undefined) {
-    if (repeatEnd) {
-      const repeatValue = repetitionEndArray.find(
-        (f: any) => f.value === repeatEnd
-      )?.unit;
-      if (repeatValue === 'onDate') {
-        repeatCount.value = undefined;
-        const date = new Date(studyStore.study.plannedEnd as string);
-        date.setHours(23, 59, 59);
-        repeatUntil.value = date;
-      } else if (repeatValue === 'after') {
-        repeatUntil.value = undefined;
-        repeatEndOption.value = repeatValue ? repeatValue : 'never';
-      } else {
-        repeatEndOption.value = 'never';
-        repeatUntil.value = undefined;
-        repeatCount.value = undefined;
-      }
-    }
-  }
-
-  function resetYearlyInterval() {
-    repeatBySetPos.value = undefined;
-    repeatByMonth.value = undefined;
-    repeatByDay.value = undefined;
-    repeatByMonthDay.value = undefined;
-    repeatUntil.value = undefined;
-  }
-  function resetRepeatEndOptions() {
-    repeatUntil.value = undefined;
-    repeatCount.value = undefined;
-  }
-  function resetRepeatFreqOptions() {
-    repeatInterval.value = undefined;
-    repeatEndOption.value = 'never';
-    resetYearlyInterval();
-    resetRepeatEndOptions();
-  }
-
-  function changeOneDayObservation() {
-    if (oneDayObservationChecked.value) {
-      if (start.value.getDate() !== end.value.getDate()) {
-        end.value = start.value;
-        end.value.setHours(
-          formerEnd.value.getHours(),
-          formerEnd.value.getMinutes(),
-          formerEnd.value.getSeconds()
+      setTimeout(() => {
+        const oldEndDate = calendarEnd.value;
+        calendarEnd.value.setTime(calendarStart.value.getTime());
+        calendarEnd.value.setHours(
+          oldEndDate.getHours(),
+          oldEndDate.getMinutes(),
+          oldEndDate.getSeconds()
         );
-        formerEnd.value = end.value;
-      }
-    } else if (!oneDayObservationChecked.value) {
-      repeatChecked.value = false;
-      resetRepeatOptions();
-      resetRepeatEndOptions();
-      resetRepeatFreqOptions();
+      }, 60);
     }
   }
 
-  function changeAllDayChecked() {
-    start.value.setHours(0, 0, 0);
-    formerStart.value = start.value;
-    end.value.setHours(23, 59, 59);
-    formerEnd.value = end.value;
-  }
-
-  const calendarInfo: Ref<string> = ref('');
-
-  function validateTime(dateType: string, dateTimeCheck?: boolean) {
-    if (dateType === 'start' && start.value > end.value) {
-      end.value = start.value;
-      formerStart.value = start.value;
-      formerEnd.value = start.value;
-      calendarInfo.value = t('scheduler.info.timeOverlapStart');
-      highlightElement('end-date');
-    } else if (dateType === 'end' && end.value < start.value) {
-      start.value = end.value;
-      formerStart.value = end.value;
-      formerEnd.value = end.value;
-      calendarInfo.value = t('scheduler.info.timeOverlapEnd');
-      highlightElement('start-time');
-    }
-
-    if (typeof dateTimeCheck === 'undefined') {
-      formerEnd.value.setHours(
-        end.value.getHours(),
-        end.value.getMinutes(),
-        end.value.getSeconds()
-      );
-      end.value = formerEnd.value;
-    } else if (dateTimeCheck && dateType === 'end') {
-      formerEnd.value = end.value;
+  function onChangeEntireDayCheckbox() {
+    if (entireDayCheckbox.value) {
+      calendarStart.value.setHours(0, 0, 0);
+      calendarEndChangedWithStart.value = true;
+      setTimeout(() => {
+        calendarEnd.value.setHours(23, 59, 59);
+      }, 60);
     }
   }
 
-  function validateDate(dateType: string) {
-    if (dateType === 'start') {
-      start.value.setHours(
-        formerStart.value.getHours(),
-        formerStart.value.getMinutes(),
-        formerStart.value.getSeconds()
-      );
-      if (start.value < minDate || start.value > maxDate) {
-        start.value = formerStart.value;
-        if (!oneDayObservationChecked.value) {
-          calendarInfo.value = t('scheduler.info.startDateNotWithinRange');
-        } else {
-          formerStart.value = start.value;
-        }
-        highlightElement('start-date');
-      } else if (start.value.getDate() > end.value.getDate()) {
-        end.value = start.value;
-        formerStart.value = start.value;
-        formerEnd.value = end.value;
-        if (!oneDayObservationChecked.value) {
-          calendarInfo.value = t('scheduler.info.dateOverlapStart');
-        }
-        highlightElement('end-date');
-      }
-    } else if (dateType === 'end') {
-      end.value.setHours(
-        formerEnd.value.getHours(),
-        formerEnd.value.getMinutes(),
-        formerEnd.value.getMinutes()
-      );
-      if (end.value < minDate || end.value > maxDate) {
-        end.value = formerEnd.value;
-        validateTime('end', true);
-        calendarInfo.value = t('scheduler.info.endDateNotWithinRange');
-        highlightElement('end-date');
-      } else if (end.value.getDate() < start.value.getDate()) {
-        end.value = start.value;
-        end.value.setHours(
-          formerEnd.value.getHours(),
-          formerEnd.value.getMinutes(),
-          formerEnd.value.getSeconds()
+  const errors: Ref<Array<MoreTableChoice>> = ref([]);
+  const warnings: Ref<Array<MoreTableChoice>> = ref([]);
+  function getErrorOrWarning(
+    type: string,
+    label: string
+  ): string | null | undefined {
+    switch (type) {
+      case 'warning': {
+        const item = warnings.value.find((el) =>
+          el.label === label ? el.value : ''
         );
-        formerEnd.value = start.value;
-        validateTime('end', true);
-        calendarInfo.value = t('scheduler.info.dateOverlapEnd');
-        highlightElement('start-date');
-      } else {
-        formerEnd.value = end.value;
-        validateTime('end', true);
+        return item?.value;
       }
-    } else {
-      calendarInfo.value = '';
-      calendarInputErrors.value.start = '';
-      calendarInputErrors.value.end = '';
+      case 'error': {
+        const item = errors.value.find((el) =>
+          el.label === label ? el.value : ''
+        );
+        return item?.value;
+      }
+      default:
+        return '';
     }
   }
 
-  function highlightElement(elClass: string) {
-    const items = document.getElementsByClassName(elClass);
-    setTimeout(() => {
-      for (const el of items) {
-        el.classList.add('highlight');
-      }
-    }, 200);
+  function validateDate(dateType: string, newValue: Date, oldValue: Date) {
+    errors.value = [];
+    warnings.value = [];
+    switch (dateType) {
+      case 'calendarStart':
+        {
+          if (newValue > calendarEnd.value) {
+            calendarStart.value = newValue;
+            if (oldValue.getDate() !== newValue.getDate()) {
+              calendarStart.value.setHours(
+                oldValue.getHours(),
+                oldValue.getMinutes(),
+                oldValue.getSeconds()
+              );
+            }
+            calendarEndChangedWithStart.value = true;
+            calendarEnd.value = calendarStart.value;
+            setTimeout(() => {
+              calendarEndChangedWithStart.value = false;
+            }, 60);
 
-    setTimeout(() => {
-      for (const el of items) {
-        el.classList.remove('highlight');
+            warnings.value.push({
+              label: 'endWasChangedWithStartValue',
+              value: t(
+                'scheduler.warningsAndErrors.endWasChangedWithStartValue'
+              ),
+            });
+          } else {
+            if (oldValue.getDate() !== newValue.getDate()) {
+              calendarStart.value.setHours(
+                oldValue.getHours(),
+                oldValue.getMinutes(),
+                oldValue.getSeconds()
+              );
+            }
+          }
+        }
+        break;
+      case 'calendarEnd': {
+        if (oldValue.getDate() !== newValue.getDate()) {
+          calendarEnd.value.setHours(
+            oldValue.getHours(),
+            oldValue.getMinutes(),
+            oldValue.getSeconds()
+          );
+        }
       }
-    }, 1000);
-  }
-
-  function repeatCheckedData() {
-    if (!repeatChecked.value) {
-      resetRepeatOptions();
     }
   }
 
-  function hasDateRangeError() {
-    const e1 = checkStartDateError(start.value);
-    const e2 = checkEndDateError(end.value);
-    return e1 || e2;
+  function toggleHasRruleValue(rruleCheckbox: boolean) {
+    hasRruleValue.value = rruleCheckbox;
+  }
+
+  function onRruleUpdate(schedulerRrule: RecurrenceRule) {
+    if (hasRruleValue.value) {
+      returnSchedule.value.rrule = schedulerRrule;
+    }
+  }
+
+  const rruleErrors: Ref<MoreTableChoice[]> = ref([]);
+  function onRruleErrorUpdate(updatedRruleErrors: MoreTableChoice[]) {
+    rruleErrors.value = updatedRruleErrors;
   }
 
   function save() {
-    if (hasDateRangeError()) {
-      return;
+    // setCalendarStart and calendarEnd into returnSchedule object
+    returnSchedule.value.dtstart = calendarStart.value.toISOString();
+    returnSchedule.value.dtend = calendarEnd.value.toISOString();
+
+    // check if repeateChecked is on -> if not set rrule: undefined, if it is save rrule with close
+    if (!hasRruleValue.value) {
+      returnSchedule.value.rrule = undefined;
     }
 
-    const s = start.value;
-    const e = end.value;
-
-    s.setMilliseconds(0);
-    s.setSeconds(0);
-    e.setMilliseconds(0);
-    e.setSeconds(0);
-
-    if (allDayChecked.value) {
-      s.setHours(0, 0, 0);
-      e.setHours(23, 59, 59);
-    }
-
-    if (repeatCount.value && repeatByDay.value?.length) {
-      repeatCount.value = (
-        parseInt(repeatCount.value) * repeatByDay.value?.length
-      ).toString();
-    }
-    if (!calendarInputErrors.value.start && !calendarInputErrors.value.end) {
+    if (!errors.value.length && !rruleErrors.value.length) {
       try {
-        const returnEvent: Event = {
-          type: 'Event',
-          dtstart: s.toISOString(),
-          dtend: e.toISOString(),
-          rrule: undefined,
-        };
-
-        if (repeatFreq.value) {
-          returnEvent.rrule = {
-            freq: repeatFreq.value,
-            until: repeatUntil.value
-              ? repeatUntil.value?.toISOString()
-              : undefined,
-            count: repeatCount.value ? parseInt(repeatCount.value) : undefined,
-            interval: repeatInterval.value,
-            byday: repeatByDay.value,
-            bymonth: repeatByMonth.value,
-            bymonthday: repeatByMonthDay.value
-              ? parseInt(repeatByMonthDay.value)
-              : undefined,
-            bysetpos: repeatBySetPos.value,
-          };
-          if (
-            !returnEvent.rrule.count &&
-            !returnEvent.rrule.until &&
-            studyStore.study.plannedEnd
-          ) {
-            const endDate: Date = new Date(studyStore.study.plannedEnd);
-            endDate.setHours(23, 59, 59);
-            returnEvent.rrule.until = endDate.toISOString();
-          }
-        }
-        dialogRef.value.close(returnEvent);
+        dialogRef.value.close(returnSchedule.value);
       } catch (e) {
         console.error('Cannot send schedule event ', e);
       }
@@ -504,55 +211,16 @@ Licensed under the Elastic License 2.0. */
   function cancel() {
     dialogRef.value.close();
   }
-
-  function resetRepeatOptions() {
-    repeatBySetPos.value = undefined;
-    repeatByMonth.value = undefined;
-    repeatByDay.value = undefined;
-    repeatByMonthDay.value = undefined;
-    repeatUntil.value = undefined;
-    repeatCount.value = undefined;
-    repeatInterval.value = undefined;
-    repeatEndOption.value = 'never';
-    repeatCountLabel.value = undefined;
-    repeatFreq.value = undefined;
-  }
-
-  const calendarInputErrors: Ref<any> = ref({
-    start: '',
-    end: '',
-  });
-
-  function checkEndDateError(endDate: Date): boolean {
-    if (endDate < minDate || endDate > maxDate) {
-      calendarInputErrors.value.end =
-        'Please enter a end date-time that lies inside the study range and is not in the past.';
-      return true;
-    } else {
-      calendarInputErrors.value.end = '';
-      return false;
-    }
-  }
-
-  function checkStartDateError(startDate: Date): boolean {
-    if (startDate < minDate || startDate > maxDate) {
-      calendarInputErrors.value.start =
-        'Please enter a start date-time that lies inside the study range.';
-      return true;
-    } else {
-      calendarInputErrors.value.start = '';
-      return false;
-    }
-  }
 </script>
 
 <template>
   <div class="scheduler relative">
     <div class="grid grid-cols-6 items-center gap-4">
-      <h6 class="col-span-6">{{ $t('scheduler.labels.event.setDates') }}</h6>
-      <div class="col-span-6 mb-1">
-        {{ $t('scheduler.labels.event.setDatesDesc') }}
+      <div class="header-info col-span-6 mb-4">
+        <h6>{{ $t('scheduler.labels.event.setDates') }}</h6>
+        <div>{{ $t('scheduler.labels.event.setDatesDesc') }}</div>
       </div>
+
       <div class="col-span-6 mb-4">
         <h5 class="text-color mb-1 font-bold">
           {{ $t('cronSchedule.example.title') }}
@@ -595,266 +263,120 @@ Licensed under the Elastic License 2.0. */
           </div>
         </div>
       </div>
-      <div v-if="!oneDayObservationChecked" class="col-span-5 font-bold">
-        {{ $t('scheduler.labels.timeframe') }}
-      </div>
+
+      <hr class="col-span-6" />
+
+      <h6 class="col-span-6">
+        {{ $t('scheduler.preview.title.individualEvent') }}
+      </h6>
+
       <div class="col-span-1 col-start-1 font-bold">
-        <span v-if="oneDayObservationChecked">{{
-          $t('scheduler.labels.timeframe')
-        }}</span>
-        <span v-else>{{ $t('global.labels.start') }}</span>
+        {{
+          singleDayEventCheckbox
+            ? t('scheduler.labels.timeframe')
+            : t('global.labels.start')
+        }}
       </div>
+
       <Calendar
-        v-model="start"
+        v-model="calendarStart"
         date-format="dd/mm/yy"
         :placeholder="'dd/mm/yyyy'"
-        :min-date="minDate"
-        :max-date="maxDate"
+        :min-date="new Date(studyStore.study.plannedStart as string)"
+        :max-date="new Date(studyStore.study.plannedEnd as string)"
+        :manual-input="false"
         autocomplete="off"
-        style="width: 100%"
-        class="start-date col-span-2 col-start-2"
-        :class="calendarInputErrors.start ? 'input-error' : ''"
-        @date-select="validateDate('start')"
-        @blur="validateDate('start')"
+        class="start-date col-span-2 col-start-2 w-full"
       />
+
       <Calendar
-        v-if="!allDayChecked"
-        v-model="start"
+        v-if="!entireDayCheckbox"
+        v-model="calendarStart"
+        :min-date="new Date(studyStore.study.plannedStart as string)"
+        :max-date="new Date(studyStore.study.plannedEnd as string)"
+        :manual-input="false"
         :placeholder="'hh:mm'"
         class="p-calendar-timeonly start-date start-time col-span-1"
         time-only
-        @blur="validateTime('start')"
       />
-      <div v-if="oneDayObservationChecked && !allDayChecked" class="col-span-2">
-        <span class="ml-1 mr-4">-</span>
-        <Calendar
-          v-model="end"
-          :placeholder="'hh:mm'"
-          class="p-calendar-timeonly end-date end-time col-span-1"
-          time-only
-          @blur="validateTime('end')"
-        />
-      </div>
 
       <div
-        v-if="calendarInputErrors.start"
-        class="error col-span-5 col-start-2"
-      >
-        {{ calendarInputErrors.start }}
-      </div>
-      <div
-        v-if="!oneDayObservationChecked"
+        v-if="!singleDayEventCheckbox"
         class="col-span-1 col-start-1 font-bold"
       >
-        {{ $t('global.labels.end') }}
+        {{ t('global.labels.end') }}
       </div>
+
       <Calendar
-        v-if="!oneDayObservationChecked"
-        v-model="end"
+        v-if="!singleDayEventCheckbox"
+        v-model="calendarEnd"
+        :min-date="calendarStart"
+        :max-date="new Date(studyStore.study.plannedEnd as string)"
+        :manual-input="false"
         date-format="dd/mm/yy"
         :placeholder="'dd/mm/yyyy'"
-        :min-date="start > minDate ? start : minDate"
-        :max-date="maxDate"
         autocomplete="off"
-        style="width: 100%"
-        class="end-date col-span-2 col-start-2"
-        :class="calendarInputErrors.end ? 'input-error' : ''"
-        @date-select="validateDate('end')"
-        @blur="validateDate('end')"
+        class="start-date col-span-2 col-start-2 w-full"
+        :class="
+          getErrorOrWarning('warning', 'endWasChangedWithStartValue')
+            ? 'calendar-warning'
+            : ''
+        "
       />
+
       <Calendar
-        v-if="!allDayChecked && !oneDayObservationChecked"
-        v-model="end"
+        v-if="!entireDayCheckbox"
+        v-model="calendarEnd"
+        :manual-input="false"
+        :min-date="new Date(studyStore.study.plannedStart as string)"
+        :max-date="new Date(studyStore.study.plannedEnd as string)"
         :placeholder="'hh:mm'"
-        class="p-calendar-timeonly end-date end-time col-span-1"
+        class="p-calendar-timeonly start-date start-time col-span-1"
+        :class="
+          getErrorOrWarning('warning', 'endWasChangedWithStartValue')
+            ? 'calendar-warning'
+            : ''
+        "
         time-only
-        @blur="validateTime('end')"
       />
-      <div v-if="calendarInfo" class="col-span-5 col-start-2">
-        <span class="pi pi-info-circle color-primary mr-2" />{{ calendarInfo }}
-      </div>
-      <div v-if="calendarInputErrors.end" class="error col-span-5 col-start-2">
-        {{ calendarInputErrors.end }}
-      </div>
-
-      <div class="col-span-2 col-start-2">
-        {{ $t('scheduler.labels.event.oneDayObservation') }}:
-        <Checkbox
-          v-model="oneDayObservationChecked"
-          class="ml-2"
-          :binary="true"
-          @change="changeOneDayObservation()"
-        />
-      </div>
-
-      <div class="col-span-2">
-        <span>{{ $t('scheduler.labels.event.allDay') }}</span>
-        <Checkbox
-          v-model="allDayChecked"
-          class="ml-2"
-          :binary="true"
-          @change="changeAllDayChecked()"
-        />
-      </div>
-
       <div
-        v-if="oneDayObservationChecked"
-        class="col-span-6 col-start-1 grid grid-cols-6 gap-4"
+        v-if="getErrorOrWarning('warning', 'endWasChangedWithStartValue')"
+        class="warning col-span-6 col-start-2"
       >
-        <hr class="col-start-0 col-span-6 mb-4 mt-4" />
-        <h6 class="col-span-6">{{ $t('scheduler.labels.event.repeat') }}</h6>
-        <div class="col-span-6 mb-2">
-          {{ $t('scheduler.labels.event.repeatDesc') }}
-        </div>
-        <div class="col-span-6 mt-4 grid grid-cols-6 items-center gap-4">
-          <div class="col-span-1">
-            {{ $t('scheduler.labels.repeat') }}:
-            <Checkbox
-              v-model="repeatChecked"
-              class="ml-2"
-              :binary="true"
-              @change="repeatCheckedData()"
-            />
-          </div>
-          <!-- Frequency: never to yearly -->
-          <div
-            v-if="repeatChecked"
-            class="col-span-5 col-start-2 grid grid-cols-5 gap-4"
-          >
-            <SelectButton
-              v-model="repeatFreq"
-              :options="repeatFreqArray"
-              option-label="label"
-              option-value="value"
-              class="col-span-5 w-full"
-              @click="setRepeatCountLabel(repeatFreq)"
-              @change="resetRepeatFreqOptions"
-            />
-          </div>
+        {{ getErrorOrWarning('warning', 'endWasChangedWithStartValue') }}
+      </div>
 
-          <!-- weekday select -->
-          <div
-            v-if="repeatChecked && repeatFreq === Frequency.Weekly"
-            class="col-span-5 col-start-2 grid grid-cols-5 gap-4"
-          >
-            <SelectButton
-              v-model="repeatByDay"
-              :options="repeatWeekdayArray"
-              option-label="label"
-              option-value="value"
-              class="col-span-5 w-full"
-              :multiple="true"
-            ></SelectButton>
-          </div>
-
-          <!-- monthly/yearly select -->
-          <div
-            v-if="
-              (repeatChecked && repeatFreq === Frequency.Yearly) ||
-              (repeatChecked && repeatFreq === Frequency.Monthly)
-            "
-            class="col-span-5 col-start-2 grid grid-cols-5 gap-4"
-          >
-            <Dropdown
-              v-model="repeatYearOption"
-              :options="repeatYearOptionArray"
-              :option-label="'label'"
-              :option-value="'value'"
-              class="col-span-3"
-              @change="resetYearlyInterval()"
-            >
-            </Dropdown>
-
-            <div
-              v-if="repeatYearOption === repeatYearOptionArray[0].value"
-              class="col-start-0 col-span-5 grid grid-cols-3 items-center gap-4"
-            >
-              <Dropdown
-                v-if="repeatFreq === Frequency.Yearly"
-                v-model="repeatByMonth"
-                :options="repeatByMonthOptionArray"
-                :option-label="'label'"
-                :option-value="'value'"
-                :placeholder="$t('scheduler.placeholder.chooseMonth')"
-              />
-              <InputText
-                v-model="repeatByMonthDay"
-                :placeholder="$t('scheduler.placeholder.enterDayOfMonth')"
-              />
-            </div>
-            <div
-              v-if="repeatYearOption === repeatYearOptionArray[1].value"
-              class="col-start-0 col-span-5 grid grid-cols-3 gap-4"
-            >
-              <Dropdown
-                v-model="repeatBySetPos"
-                :options="repeatBySetPosOptionArray"
-                :option-label="'label'"
-                :option-value="'value'"
-                :placeholder="$t('scheduler.placeholder.chooseGeneralInterval')"
-              />
-              <Dropdown
-                v-model="repeatByDay"
-                :options="repeatByDayForYearArray"
-                :option-label="'label'"
-                :option-value="'value'"
-                :placeholder="$t('scheduler.placeholder.chooseDayInterval')"
-              />
-              <Dropdown
-                v-if="repeatFreq === Frequency.Yearly"
-                v-model="repeatByMonth"
-                :options="repeatByMonthOptionArray"
-                :option-label="'label'"
-                :option-value="'value'"
-                :placeholder="$t('scheduler.placeholder.chooseMonth')"
-              />
-            </div>
-          </div>
-
-          <hr
-            v-if="repeatFreq === Frequency.Weekly"
-            class="col-start-0 col-span-6 mb-4 mt-4"
+      <div class="checkboxen col-span-7 col-start-2 grid grid-cols-2">
+        <div class="col-start-1">
+          {{ $t('scheduler.labels.event.oneDayObservation') }}:
+          <Checkbox
+            v-model="singleDayEventCheckbox"
+            class="ml-2"
+            :binary="true"
+            @change="onChangeSingleDayEventCheckbox()"
           />
-          <div
-            v-if="repeatChecked && repeatFreq"
-            class="col-start-0 col-span-1 self-center font-medium"
-          >
-            {{ $t('scheduler.labels.repetitionEnd') }}
-          </div>
-          <div
-            v-if="repeatChecked && repeatFreq"
-            class="col-span-5 col-start-2 gap-4"
-          >
-            <SelectButton
-              v-model="repeatEndOption"
-              :options="repetitionEndArray"
-              option-label="label"
-              option-value="value"
-              class="mr-4 inline w-full"
-              @click="setRepetitionEnd(repeatEndOption)"
-            ></SelectButton>
-            <div v-if="repeatEndOption === 'after'" class="inline">
-              <InputText
-                v-model="repeatCount"
-                type="text"
-                :placeholder="$t('scheduler.placeholder.enterRepeatCount')"
-              />
-              <span class="ml-2">{{ repeatCountLabel }}</span>
-            </div>
-            <div v-else-if="repeatEndOption === 'onDate'" class="col-span-2">
-              <Calendar
-                v-show="false"
-                v-model="repeatUntil"
-                placeholder="dd/mm/yyyy"
-                date-format="dd/mm/yy"
-                autocomplete="off"
-                style="width: 100%"
-                :class="'col-span-5'"
-              />
-            </div>
-          </div>
+        </div>
+        <div>
+          <span>{{ $t('scheduler.labels.event.allDay') }}</span>
+          <Checkbox
+            v-model="entireDayCheckbox"
+            class="ml-2"
+            :binary="true"
+            @change="onChangeEntireDayCheckbox()"
+          />
         </div>
       </div>
+
+      <!-- import repeat -->
+      <AbsoluteSchedulerRepetition
+        v-if="singleDayEventCheckbox"
+        :rrule-checked="hasRruleValue"
+        :scheduler-rrule="returnSchedule.rrule"
+        class="col-span-8"
+        @on-rrule-checkbox-change="toggleHasRruleValue($event)"
+        @on-rrule-change="onRruleUpdate($event)"
+        @on-rrule-error="onRruleErrorUpdate($event)"
+      />
     </div>
 
     <div style="height: 100px"></div>
@@ -908,5 +430,9 @@ Licensed under the Elastic License 2.0. */
     padding: 0.5rem;
     border-radius: 6px;
     background-color: var(--surface-50);
+  }
+
+  :deep(.calendar-warning) {
+    border: var(--border-weight) var(--border-style) var(--yellow-600);
   }
 </style>
